+++ R/gg_col.R
@@ -1242,7 +1242,7 @@ gg_col <- function(
   }
   else if (col_legend_place == "none") {
     plot <- plot +
-      ggplot2::theme(legend.position = col_legend_place)
+      ggplot2::guides(col = FALSE, fill = FALSE)
   }
 
   #remove gridlines as per x_gridlines and y_gridlines. Guess if NULL