+++ R/gg_sf.R
@@ -1110,9 +1110,9 @@ gg_sf <- function(
       col_legend_place <- "bottom"
     }
     else if (#(identical(rlang::eval_tidy(col, data), rlang::eval_tidy(x, data))) |
-             #(identical(rlang::eval_tidy(col, data), rlang::eval_tidy(y, data))) |
-             (identical(rlang::eval_tidy(col, data), rlang::eval_tidy(facet, data))) |
-             (identical(rlang::eval_tidy(col, data), rlang::eval_tidy(facet2, data)))) {
+      #(identical(rlang::eval_tidy(col, data), rlang::eval_tidy(y, data))) |
+      (identical(rlang::eval_tidy(col, data), rlang::eval_tidy(facet, data))) |
+      (identical(rlang::eval_tidy(col, data), rlang::eval_tidy(facet2, data)))) {
       col_legend_place <- "none"
     }
     else if (col_numeric | col_date | col_datetime | col_time) col_legend_place <- "right"
@@ -1151,7 +1151,7 @@ gg_sf <- function(
   }
   else if (col_legend_place == "none") {
     plot <- plot +
-      ggplot2::theme(legend.position = col_legend_place)
+      ggplot2::guides(col = FALSE, fill = FALSE)
   }
 
   #remove gridlines as per x_gridlines and y_gridlines. Guess if NULL