+++ R/gg_polygon.R
@@ -57,6 +57,7 @@
 #' @param facet_scales Whether facet scales should be "fixed" across facets, "free" in both directions, or free in just one direction (i.e. "free_x" or "free_y"). Defaults to "fixed".
 #' @param facet_space Whether facet space should be "fixed" across facets, "free" to be proportional in both directions, or free to be proportional in just one direction (i.e. "free_x" or "free_y"). Defaults to "fixed". Only applies where the facet layout is "grid" and facet scales are not "fixed".
 #' @param facet_layout Whether the layout is to be "wrap" or "grid". If NULL and a single facet (or facet2) argument is provided, then defaults to "wrap". If NULL and both facet and facet2 arguments are provided, defaults to "grid".
+#' @param facet_switch Whether the facet layout is "grid", whether to switch the facet labels to the opposite side of the plot. Either "x", "y" or "both".
 #' @param titles A function to format the x, y and col titles. Defaults to snakecase::to_sentence_case.
 #' @param caption Caption title string.
 #' @param theme A ggplot2 theme.
@@ -148,6 +149,7 @@ gg_polygon <- function(
     facet_scales = "fixed",
     facet_space = "fixed",
     facet_layout = NULL,
+    facet_switch = NULL,
     caption = NULL,
     titles = snakecase::to_sentence_case,
     theme = NULL) {
@@ -410,7 +412,7 @@ gg_polygon <- function(
   else if (facet_layout == "grid") {
     if (!facet_null & !facet2_null) {
       plot <- plot +
-        ggplot2::facet_grid(
+        ggplot2::facet_grid(switch = facet_switch,
           rows = ggplot2::vars(!!facet2),
           cols = ggplot2::vars(!!facet),
           scales = facet_scales,
@@ -420,7 +422,7 @@ gg_polygon <- function(
     }
     else if (!facet_null & facet2_null) {
       plot <- plot +
-        ggplot2::facet_grid(
+        ggplot2::facet_grid(switch = facet_switch,
           cols = ggplot2::vars(!!facet),
           scales = facet_scales,
           space = facet_space,
@@ -429,7 +431,7 @@ gg_polygon <- function(
     }
     else if (facet_null & !facet2_null) {
       plot <- plot +
-        ggplot2::facet_grid(
+        ggplot2::facet_grid(switch = facet_switch,
           rows = ggplot2::vars(!!facet2),
           scales = facet_scales,
           space = facet_space,