+++ NEWS.md
@@ -2,8 +2,8 @@
 
 * Breaking: removed gg_function.
 * Improved gg_sf default alpha.
-* Supported gg_sf working on non-standard named geometry.
-* Adjust theme defaults.
+* Supported gg_sf to work with non-standard named geometry.
+* Adjusted theme defaults (e.g. title positioning).
 
 # ggblanket 2.0.0