+++ NEWS.md
@@ -1,11 +1,12 @@
 # ggblanket 2.0.9000
 
-* Breaking: redesigned the themes, including their arguments. 
-* New theme functions: `gg_theme_light`, `gg_theme_dark` and `gg_theme_void`.
+* New theme functions: `light_mode` and `dark_mode`.
+* Breaking: removed `gg_theme`. 
 * Breaking: renamed `*_grid` arguments to `*_gridlines`.
 * Breaking: removed `void` argument.
 * Breaking: removed `gg_function`.
 * Breaking: made `col_rescale` require a function (e.g. `scales::rescale()`).
+* For raster, removed col aesthetic so that legend would work.
 * Improved `gg_sf` default `alpha`.
 * Supported `gg_sf` to work with non-standard named geometry.