+++ NEWS.md
@@ -1,4 +1,4 @@
-# ggblanket 2.0.9000
+# ggblanket 3.0.9000
 
 * New theme functions: `light_mode` and `dark_mode`.
 * Breaking: removed `gg_theme`.